+++ docs/hook/useCountDown/demos/demo1.tsx
@@ -4,16 +4,11 @@ import { Button, Input } from 'antd';
 import 'antd/lib/input/style/css';
 
 export default () => {
-  const [{ isCountDowning, remaning }, { start }] = useCountDown();
+  const [{ isCountDowning, remaining }, { start }] = useCountDown();
   return (
     <div>
-      <Input suffix={
-            isCountDowning ? (
-              <span>
-                {`${remaning}s`}
-              </span>
-            ) : <Button onClick={() => start()}>获取验证码</Button>
-        }
+      <Input
+        suffix={isCountDowning ? <span>{`${remaining}s`}</span> : <Button onClick={() => start()}>获取验证码</Button>}
       />
     </div>
   );